+++ src/features/viewer/components/content-preview.tsx
